Adjusted sphere case to match experiment
Sphere is now falling and spinning in z-axis. Adjusted stat and pvsm
for the altered test condition. The spin is fixed at 2.4U/D and
Re=2100, which are the value measured in the experiment. Using tanh
drop speed, but the results are similar as when using linear ramping.
Added mpi code to speed things up.