Added motion to square project
Added a simple spring-mass motion prediction scheme to lotus. This
is absurdly easy in lotus (soooo much cleaner than the old BDIM).
The position is integrated like usual, the velocity uses an
estimate of the change in acceleration. This is not as nice as
a predictor correct for the velocity, but it avoids the correction.

I added the A* and A*[10] summary statistics to stat.R. This is pretty
easy using dplyr with filter, lag, and lead. To get the top 10% I used
cume_dist.