Update flag project
The beam equation is forth order and (it would appear) that means it
MUST be solved using real(8).

Added order of accurate flag in stencil. Added modeshapes.py to solve
for the exact wavenumbers of the fundamental modes and also to test
the influence of the boundary stencil order of accuracy. 3rd order is
better for the clamped end, but 2nd gives lower error for the free end.

Implemented these changed in lotus as well as printing out the edge
deflection in fort.11 and plotting it in stat. Everything with the
beam looks good now.