Add force_check project
New project to help validate and debug force calculation routines.

Current test case is a sphere (defined with the SDF) moving in x,y,z immersed in a pressure field that increases linearly in x,y,z. The pressure integral should therefore give the volume.

Current result is that the mean volume is  overestimated: measuring force= V(R+1.3h) instead of V(R). Something is off with the curvature correction.

There is also small measurement noise due to motion through the grid. It's amplitude is around .2% of the surface area. This is acceptable and likely unavoidable.